CV-589 Refactoring
